# docs/api/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Arduino LED Controller API v2",
    "version": "5.0.0-alpha.2",
    "description": "Moduláris HTTP API az Arduino LED Controller szerverhez."
  },
  "servers": [
    {
      "url": "/",
      "description": "Aktuális szerver"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "led_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
      "ApiSuccess": {
        "type": "object",
        "required": [
          "success",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "data": {},
          "meta": {
            "type": "object"
          }
        }
      },
...
      "ManagedApiTokenCreate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 96
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ManagedApiTokenUpdate": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 96
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "operator",
              "viewer"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Success": {
        "description": "Sikeres válasz",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiSuccess"
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v2/firmware/status": {
      "get": {
        "summary": "Firmware állapot",
        "tags": [
          "Firmware"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/v2/tokens": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "API-tokenek listázása titkok nélkül",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Új forgatható API-token létrehozása",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedApiTokenCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/api/v2/tokens/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "tags": [
          "Tokens"
        ],
        "summary": "Forgatható API-token módosítása",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedApiTokenUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Forgatható API-token törlése",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
    "/api/v2/tokens/{id}/actions/rotate": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "API-token rotálása és a régi token letiltása",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedApiTokenUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          }
        }
      }
    },
...
  }
}